Heavily change the way initramfs is done
Stop building an initramfs that is separate from the root
but instead basically make the system an initramfs based
system.  If there is a root device with /sbin/init on it,
it will execute it. Otherwise it will run from ram.

By doing this, there is no longer a need to determine
binary dependencies.  The initramfs is a very useful system.

Also, include a 'lxc-is-container' script that exits
with "not lxc".  That will get overwritten by etc/init/rc.sysinit
to have the correct information.