return filesystem back to ext3
using ext3 means that online resizing can be done.

We could instead of doing online resizing invoke resize2fs
from the ramdisk, but then there would be no way for the
user to turn it on at that point (as user-data is not available).

Due to speed of the resize we dont do resize by default.