mount 'ephemeral0' to /mnt
This will only really work in nova, but if there is a disk named 'ephermal0',
then mount it to /mnt.  This also only works if a datasource is found (as it
runs in cirros-apply).

Also, note that previously, in 'ramdisk only' mode, we would mount all devices
into /media.  That function has been dropped.

LP: #1022619