fix growing gpt partitions by updating sfdisk in util-linux to 2.27
growpart will not use sfdisk to do gpt resizing if the version is
less than 2.26.3.  That is because of some bugs in sfdisk (LP: #1474090).

This cherry-picks the buildroot changes to get to 2.27.  We stop there
only because it is sufficient, and it looks like there might be further
changes required to update buildroot further.

LP: #1605589