make disk image (disk.img) contain only /boot.
This removes everything except /boot from a disk.img.
On first boot of the disk image, the loaded initramfs in /boot
will copy itself to the target.  Since /boot is not in the initramfs
there is no problem copying to the blank target.

The result of this is a disk1.img file that is much smaller.
for comparison, the output of this local build and the 20160722 daily
for x86_64:
   disk.img 17961472 -> 12846080 (29% saved)
   rootfs.img.gz 16744808 -> 11899176 (29% saved)

If curtin's src/init doesn't find a root filesystem (no /sbin/init) but
does find a filesystem with the right label (cirros-rootfs), then it
copies itself to the target and continues booting as if it were always
there.

src/init now uses search_for_root and search_for_blank to find root
filesystem and blank disk.  It would be nice to move the "install"
hunk out of src/init also.  We need to improve this too, to deal with
existing files overwriting and such.