src/etc/modules: config_drive cdrom support.
config_drive can present as a SCSI CDROM, so we need suitable modules
loaded to access it.

Added in a new section for block storage.

There is no growth in size on Ubuntu, because all 3 modules are
compiled-in already, but the module loading is useful for other kernels
where they might be modular.

LP: #1715009
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>