Add stream data and creation of metadata for lxd and -lxc.tar.xz.
This adds a lxd.tar.xz metadata file and also the information in the
stream output so that cirros can be utilized by lxd.  As the needed
files (-lxc.tar.xz and -lxd.tar.xz) are purely new content, the change to
mirror-dump-sstream-data allows us to create them for old releases.

A small re-factor in mirror-dump-sstream-data was required to more
easily create missing -lxd.tar.gz and -lxd.tar.xz and get the
data to create the combined_rootxz_sha256 information.

Once this is published, we'll be able to do:
  lxc remote add --public --protocol=simplestreams cirros \
       https://download.cirros-cloud.net/
  lxc launch cirros:0.4.0

and also:
  lxc launch cirros:0.4.0 "--config=user.user-data=$(cat user-data.txt)"