Enable build on Ubuntu 16.04.3 LTS
This patch enables building CirrOS on top of Ubuntu 16.04.3 LTS.

The build failed on Xenial because GCC 5.x changed the preprocessor
output, which ncurses build scripts relied on. This has been reported and
fixed in ncurses and buildroot:

  https://trac.sagemath.org/ticket/18301
  https://git.busybox.net/buildroot/commit/?id=4297e7298edbef6dde75d8bf468441c6a242da5c

However, the fix applied on buildroot does not match the version string
`5.4.0`, returned by `gcc -dumpversion` on Xenial. With this change, the
gcc version `5.4.0` is matched and ncurses preprocessor flags are set
properly.

It was necessary to add python package to bin/system-setup because python
is not installed by default on Xenial.