grab-kernels: support newer kernel packages.
Kernel packages and naming schemes between release 16.04 (kernel 4.4)
and 18.04 (kernel 4.15).  Then we noticed in 4.4.0-148 that
the 4.4 packages had changed to similar format as 4.15, but with
different availability.

The issue with 4.15 kernels in Ubuntu is that they do not provide
any powerpc (ppc64 big endian) kernels.  So before we could move
there we would need to find ppc64 kernel or drop powerpc and ppc64
arches (ppc64el is fine).