load-modules: try to check is module built-in into kernel
Image contains a list of kernel modules to load. Several of them are
built-in into kernel so 'modprobe' fails and complains that kernel
module is not found.

To make it more quiet let first check in /sys/module/ directory is
wanted module already a part of kernel. Still some failures will be
present as some modules are not listed there (vfat, nls_* ones).

Closes: #23