dropbear: Handle '/etc/dropbear' as a symlink
New dropbear versions have '/etc/dropbear' as a symlink to
'/var/run/dropbear'. Furthermore, this symlink seems to be
on a read-only mount, since the init script is not able to
really delete it, and it's there after booting. Only after
the boot process is complete one can remove it.

The result is that one cannot log in with ssh to a clean
CirrOS instance.

To handle this situation properly we need to operate on the
symlink target - '/var/run/dropbear'- in the init script.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>