Add the 'none' datasource.
The 'none' datasource will just be found no matter what.
It allows easily / intentionally using without a datasource.

A change in behavior of cirros-ds was needed as it previously
took 'dslist=none' as a alias for 'dslist='.  Explicitly setting
dslist= is inteneded to read the value from /etc/cirros-ds/config.