Fix handling of /etc/dropbear
buildroot starts out with /etc/dropbear being a symlink to
/var/run/dropbear in order to allow dropbear to start even when /etc is
located on a RO filesystem. The startup script is then to check whether
the symlink can be replaced by an actual directory, which then allows to
permanently store the hostkeys that dropbear generates, ensuring the
keys will persist across reboots. This patch amends the S50dropbear init
script to copy this functionality from the original buildroot script.

Also add ed25519 to the list of hostkey types that are generated before
dropbear starts.

Fixes: #118

Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>